PT 147263283 - Workaround for issue in STS manifest editor
STS manifest editor deletes all error and warning markers on document
save. This is a problem with the YEdit editor that STS manifest editor
extends, and not with vscode manifest language server. This is just a
temporary workaround until STS no longer needs the YEdit-based editor
and uses LSP4E editor instead.