Split jdt ls extension into two parts
One pugin to contain the bulk of the code, which is independent of jdt.ls, lsp4e and lsp4j.
Second plugin to contain the jdt.ls specific DelegateCommandHandler implementation.

This should allow us to somehow re-use the bulk of the code in STS4 eclipse, i.e. when
we have access to JDT from eclipse directly, and there is no jdt.ls process around.