Moving towards more easy to re-use JMXConnector
The "create connector -> do something -> close" pattern
has now been turned into a method. This paves the way
for making management of the connectors smarter
so that, for example, a single connector could be
cached for some time and re-used for more than
one task.

Creating the connector is a slow process especially
for remote apps, so this is a bit of a performance
bottleneck and its a bad idea to create a new
connector for every small request.