Fix JavaModelException with non-Java projects (#1188)
Added a check that the JDT project uses to confirm the project is a Java
project. JavaCore#create will return an object even if the project isn't
a Java project and is therefore insufficient. Bring up the context menu
on non Java projects to the see the JME logged without this fix
(possibly other repercussions as well).